+++ client/src/App.js
@@ -7,13 +7,12 @@ import { useEffect, useState } from 'react';
 import BlogLay from './components/BlogLay/BlogLay';
 import Nav from "./components/NavBar/Navbar.js"
 import Manage from './pages/SubscribeandLogin/Manage';
+import Subscribe from './pages/SubscribeandLogin/Subscribe'
 import toast, { Toaster } from 'react-hot-toast';
 import About from "./pages/About/About"
 import Footer from './components/Footer/Footer';
 import Category from './components/Category/Category';
 import Contact from './pages/Contact/Contact';
-import Subscribe from './components/Subscribe/Subscribe'
-
 
 
 function App() {
@@ -68,11 +67,10 @@ function App() {
           <Route path="/blogs/:id" element={<Category/>} />
           <Route path="/login" element={<Login onLogin={onLogin}/>} />
           <Route path="/:id" element={<BlogLay currentUser={currentUser}/>} />
-          {/* <Route path="/subscribe" element={<Subscribe/>} /> */}
+          <Route path="/subscribe" element={<Subscribe/>} />
           <Route path="/your_account" element={<Manage currentUser={currentUser} toast={toast} Toaster={Toaster}/>} />
           </Routes>
       
-      <Subscribe/>
       <Footer/>
     </div>
   );
