+++ client/src/pages/Home.js
@@ -4,7 +4,6 @@ import Carousel from "../components/Carousel/Carousel";
 import Header from "../components/Header/Header";
 import { useState,useEffect } from "react";
 import { animateScroll as scroll } from 'react-scroll';
-
  
 
 function Home () {
